requirements.txt
cryptography
psycopg-binary
python-dotenv
